Release 1.14.05
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@1448 98333e67-4a24-44d7-a75c-e53540dd3050

+++ advanced/lib/abyss/src/socket_win.c
@@ -347,7 +347,7 @@ channelWrite(TChannel *            const channelP,
         ) {
         size_t const maxSend = (size_t)(-1) >> 1;
 
-        size_t rc;
+        int rc;
         
         rc = send(socketWinP->winsock, &buffer[len-bytesLeft],
                   MIN(maxSend, bytesLeft), 0);
